# lib/api-zod/src/generated/types/arcadeAccount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArcadeAccount {
  guildId: string;
  userId: string;
  username: string;
  wallet: number;
  bank: number;
  xp: number;
  achievementCount: number;
  dailyStreak: number;
  /** @nullable */
  lastDailyAt: string | null;
}