# lib/api-zod/src/generated/types/arcadeActionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArcadeAccount } from './arcadeAccount';

export interface ArcadeActionResult {
  ok: boolean;
  message: string;
  account: ArcadeAccount;
}