# lib/api-zod/src/generated/types/arcadeActivity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArcadeActivity {
  id: number;
  type: string;
  title: string;
  detail: string;
  /** @nullable */
  amount: number | null;
  createdAt: string;
}