# lib/api-zod/src/generated/types/arcadeMoneyMove.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArcadeMoneyMoveKind } from './arcadeMoneyMoveKind';

export interface ArcadeMoneyMove {
  kind: ArcadeMoneyMoveKind;
  amount: number;
  username: string;
}