# lib/api-zod/src/generated/types/arcadeOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArcadeAccount } from './arcadeAccount';
import type { ArcadeActivity } from './arcadeActivity';
import type { ArcadeInventoryItem } from './arcadeInventoryItem';
import type { ArcadeSettings } from './arcadeSettings';
import type { ArcadeShopItem } from './arcadeShopItem';

export interface ArcadeOverview {
  settings: ArcadeSettings;
  account: ArcadeAccount | null;
  inventory: ArcadeInventoryItem[];
  shop: ArcadeShopItem[];
  leaderboard: ArcadeAccount[];
  activity: ArcadeActivity[];
  memberCount: number;
}