# lib/api-zod/src/generated/types/arcadeSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArcadeSettingsToggles } from './arcadeSettingsToggles';

export interface ArcadeSettings {
  guildId: string;
  guildName: string;
  prefix: string;
  interestBps: number;
  dailyAmount: number;
  toggles: ArcadeSettingsToggles;
}