# lib/api-zod/src/generated/types/arcadeSettingsUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArcadeSettingsUpdateToggles } from './arcadeSettingsUpdateToggles';

export interface ArcadeSettingsUpdate {
  prefix?: string;
  interestBps?: number;
  dailyAmount?: number;
  toggles?: ArcadeSettingsUpdateToggles;
}