# lib/api-zod/src/generated/types/discordStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordStatus {
  connected: boolean;
  /** @nullable */
  botName: string | null;
  /** @nullable */
  botTag: string | null;
  guildCount: number;
  message: string;
  /** @nullable */
  primaryGuildId: string | null;
  /** @nullable */
  primaryGuildName: string | null;
}